Rename func_api to map_machinery
